+++ ReliableAncientScepter/Thunderstore/CHANGELOG.md
@@ -1,4 +1,9 @@
 <details>
+<summary>1.0.1 </summary>
+
+* Forgot to swap shaders.
+</details>
+<details>
 <summary>1.0.0 </summary>
 
 * Initial release.